+++ frontend/src/components/Landing2/Footer.js
