+++ frontend/src/pages/404.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-import Layout from "../components/landing/Layout"
+import Layout from "../components/Landing2/Layout"
 import SEO from "../components/seo"
 
 const NotFoundPage = () => (
