+++ frontend/src/pages/index.js
@@ -1,8 +1,8 @@
 import React from "react"
 
-import Layout from "../components/landing/Layout"
+import Layout from "../components/Landing2/Layout2"
 import SEO from "../components/seo"
-import Landing from "../components/landing/Landing"
+import Landing from "../components/Landing2/Landing"
 
 const IndexPage = () => (
   <Layout>
